scripts/requirements.txt
# Python requirements for Abyz-AutoRA
# Run: pip install -r requirements.txt

# Word document processing
docx-mailmerge==0.5.0
python-docx==1.1.0

# Utilities
python-dateutil==2.8.2
